+++ src/TestSuite/Traits/MockTrait.php
@@ -17,8 +17,8 @@ use Cake\Controller\ComponentRegistry;
 use Cake\Controller\Controller;
 use Cake\Datasource\ConnectionManager;
 use Cake\ORM\Entity;
-use Cake\View\View;
 use Cake\Utility\Inflector;
+use Cake\View\View;
 
 /**
  * A mock trait
@@ -84,6 +84,19 @@ trait MockTrait
         return $this->getMockBuilder($className)
             ->setMethods($methods)
             ->setConstructorArgs([new View])
+            ->getMock();
+    }
+
+    /**
+     * Mocks a shell
+     * @param string $className Shell class name
+     * @param array|null $methods The list of methods to mock
+     * @return \PHPUnit\Framework\MockObject\MockObject
+     */
+    protected function getMockForShell($className, $methods = ['_stop', 'in'])
+    {
+        return $this->getMockBuilder($className)
+            ->setMethods($methods)
             ->getMock();
     }